[mlir] move cell expansion into SCF lowering
